requirements.txt
asgiref==3.11.1
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
numpy==2.4.2
pandas==3.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pycparser==3.0
PyMuPDF==1.27.1
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
xlrd==2.0.2
